moved tests to correct package
